requirements.txt
backcall==0.1.0
bleach==2.1.3
bson==0.5.6
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
decorator==4.3.0
defusedxml==0.5.0
Django==2.0.5
django-allauth==0.36.0
django-extensions==2.0.7
django-request==1.5.4
django-rest-auth==0.9.3
djangorestframework==3.8.2
djangorestframework-jwt==1.8.0
entrypoints==0.2.3
html5lib==1.0.1
idna==2.6
ipykernel==4.8.2
ipython==6.4.0
ipython-genutils==0.2.0
ipywidgets==7.2.1
jedi==0.12.0
Jinja2==2.10
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
MarkupSafe==1.0
mistune==0.8.3
nbconvert==5.3.1
nbformat==4.4.0
nose==1.3.7
notebook==5.5.0
numpy==1.14.3
oauthlib==2.1.0
pandas==0.23.0
pandocfilters==1.4.2
parso==0.2.1
permission==0.4.1
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
psycopg2-binary==2.7.4
Pygments==2.2.0
PyJWT==1.6.4
pyotp==2.2.6
pypng==0.0.18
PyQRCode==1.2.1
python-dateutil==2.7.3
python3-openid==3.1.0
pytz==2018.4
pywinpty==0.5.3
pyzmq==17.0.0
qtconsole==4.3.1
requests==2.18.4
requests-oauthlib==1.0.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
terminado==0.8.1
testpath==0.3.1
tornado==5.0.2
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.2.1
